requirements.txt
# Knowledge Graph Platform - Python Dependencies
# Neo4j drivers and OGM
neo4j==5.25.0
py2neo==2021.2.4

# Data processing
pandas==2.2.3
pydantic==2.9.2

# NLP/Extraction
spacy==3.8.2

# API
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Visualization
matplotlib==3.9.2
networkx==3.4.2
pyyaml==6.0.2
